+++ tools/monitor/model/src/test/java/is/codion/tools/monitor/EntityServerMonitorTest.java
@@ -94,10 +94,10 @@ public class EntityServerMonitorTest {
 		clientMonitor.clientInstanceTableModel().items().refresh();
 		assertEquals(1, clientMonitor.clientInstanceTableModel().items().visible().count());
 		RemoteClient remoteClient = clientMonitor.clientInstanceTableModel().items().visible().get(0);
-		assertEquals(connectionProvider.clientId(), remoteClient.id());
+		assertEquals(connectionProvider.clientId(), remoteClient.clientId());
 		assertEquals(UNIT_TEST_USER, remoteClient.user());
 
-		clientMonitor.server().disconnect(remoteClient.id());//disconnects the client
+		clientMonitor.server().disconnect(remoteClient.clientId());//disconnects the client
 
 		clientMonitor.refresh();
 		assertEquals(0, clientMonitor.clientInstanceTableModel().items().visible().count());