+++ tools/monitor/ui/src/main/java/is/codion/tools/monitor/ui/ClientMonitorPanel.java
@@ -151,7 +151,7 @@ public final class ClientMonitorPanel extends JPanel {
 
 	private void disconnect() throws RemoteException {
 		for (RemoteClient remoteClient : model.clientInstanceTableModel().selection().items().get()) {
-			model.server().disconnect(remoteClient.id());
+			model.server().disconnect(remoteClient.clientId());
 			model.clientInstanceTableModel().items().remove(remoteClient);
 		}
 	}